move isValidEmail to shared/util.ts this deduplicates some code
