+++ src/renderer/components/dialogs/CreateChat.tsx
@@ -26,6 +26,7 @@ import { GroupImage } from './Group-Styles'
 import { DeltaDialogQrInner } from './QrInviteCode'
 import { JsonContact, DCContact } from '../../../shared/shared-types'
 import { DialogProps } from '.'
+import { isValidEmail } from '../../../shared/util'
 const { remote } = window.electron_functions
 
 export default function CreateChat(props: {
@@ -559,24 +560,4 @@ function CreateGroupInner(props: {
       )}
     </>
   )
-}
-
-export function isValidEmail(email: string) {
-  // empty string is not allowed
-  if (email === '') return false
-  const parts = email.split('@')
-  // missing @ character or more than one @ character
-  if (parts.length !== 2) return false
-  const [local, domain] = parts
-  // empty string is not valid for local part
-  if (local === '') return false
-  // domain is too short
-  if (domain.length <= 3) return false
-  const dot = domain.indexOf('.')
-  // invalid domain without a dot
-  if (dot === -1) return false
-  // invalid domain if dot is (second) last character
-  if (dot >= domain.length - 2) return false
-
-  return true
 }
