+++ src/shared/util.ts
@@ -1,3 +1,23 @@
 export function integerToHexColor(integerColor: number) {
   return '#' + integerColor.toString(16)
 }
+
+export function isValidEmail(email: string) {
+  // empty string is not allowed
+  if (email === '') return false
+  const parts = email.split('@')
+  // missing @ character or more than one @ character
+  if (parts.length !== 2) return false
+  const [local, domain] = parts
+  // empty string is not valid for local part
+  if (local === '') return false
+  // domain is too short
+  if (domain.length <= 3) return false
+  const dot = domain.indexOf('.')
+  // invalid domain without a dot
+  if (dot === -1) return false
+  // invalid domain if dot is (second) last character
+  if (dot >= domain.length - 2) return false
+
+  return true
+}
